+++ CHANGES.rst
@@ -3,7 +3,7 @@ CHANGES
 =======
 
 
-0.3.1 (unreleased)
+0.4.0 (unreleased)
 ------------------
 
 - Pass ``ignore_hidden`` from ``read_and_parse_archive()`` to
@@ -12,6 +12,12 @@ CHANGES
 - Add Excel tests for ``IF()``.
 
 - Add ``NOT()`` function.
+
+- Implemented ``BIN2OCT()``, ``BIN2DEC()``, ``BIN2HEX()``, ``OCT2BIN()``,
+  ``OCT2DEC()``, ``OCT2HEX()``, ``DEC2BIN()``, ``DEC2OCT()``, ``DEC2HEX()``,
+  ``HEX2BIN()``, ``HEX2OCT()``, ``HEX2DEC()``.
+
+- Drop Python 3.7 support.
 
 
 0.3.0 (2021-05-13)